Silence compiler warnings

+++ src/usb_libusb.c
@@ -340,7 +340,7 @@ static int usbdev_send(const union filedescriptor *fd, const unsigned char *bp, 
    * 0.
    */
   do {
-    tx_size = (mlen < fd->usb.max_xfer)? mlen: fd->usb.max_xfer;
+    tx_size = ((int) mlen < fd->usb.max_xfer)? (int) mlen: fd->usb.max_xfer;
     if (fd->usb.use_interrupt_xfer)
       rv = usb_interrupt_write(udev, fd->usb.wep, (char *)bp, tx_size, 10000);
     else
@@ -422,7 +422,7 @@ static int usbdev_recv(const union filedescriptor *fd, unsigned char *buf, size_
 	  if (usb_fill_buf(udev, fd->usb.max_xfer, fd->usb.rep, fd->usb.use_interrupt_xfer) < 0)
 	    return -1;
 	}
-      amnt = buflen - bufptr > nbytes? nbytes: buflen - bufptr;
+      amnt = buflen - bufptr > (int) nbytes? (int) nbytes: buflen - bufptr;
       memcpy(buf + i, usbbuf + bufptr, amnt);
       bufptr += amnt;
       nbytes -= amnt;
@@ -506,7 +506,7 @@ static int usbdev_recv_frame(const union filedescriptor *fd, unsigned char *buf,
 	  return -1;
 	}
 
-      if (rv <= nbytes)
+      if (rv <= (int) nbytes)
 	{
 	  memcpy (buf, usbbuf, rv);
 	  buf += rv;
